test: Run scenarios w/multiple Terraform versions
Need to be able to verify that future changes do or do not break when
used with older versions of Terraform. This commit adds multiple
platforms, one for each major/minor release of Terraform that is within
the range "supported" by the module(s).

This also removes the reporting for each test run; since there is not
any real validation of the environment (other than success/failure of
`terraform apply`) capturing the output of the test runs serves no value.