fix: Support secret variable marked sensitive
As reported by @cassus, when the `secret` input was marked as sensitive
this module would fail due to a sensitive value being used as a key to
the secret value resource. This change removes the `for_each` logic in
favour of a simple `count`, ensuring the potentially sensitive input
does not get used as a key for the resource.

Closes #51.