fix: Update random provider example
Terraform random provider generates a value that is marked as sensitive
and not fully known until applied; this caused problems for the random
provider example and led to the initial recommendation to apply in
two phases. The updated example shows how to use the module to create
an empty secret with accessors assigned, and then create a new version
with the value form the Random provider.